Now able to use launch_exp, created new yaml files
