+++ configs/exps/is2re_aux/is2re_aux_exps.yaml
@@ -1,0 +1,215 @@
+job:
+  mem: 32GB
+  cpus: 4
+  # gres: gpu:rtx8000:1
+  partition: long
+  # time: 15:00:00
+
+default:
+  mode: train
+  optim:
+    batch_size: 256
+    eval_batch_size: 256
+    optimizer: AdamW
+  cp_data_to_tmpdir: True
+  normalizer: null
+  model:
+    otf_graph: False
+    max_num_neighbors: 40
+    noisy_nodes: True
+
+  trainer: single
+  logger: wandb
+
+  task:
+    dataset: lmdb_noisy #single_point_lmdb
+    description: "Relaxed state energy prediction from initial structure."
+    type: regression
+    metric: mae
+    labels:
+      - relaxed energy
+
+runs:
+  - config: faenet-is2re_aux-all
+    note: "Cosine-very-low-lr"
+    optim:
+      lr_initial: 0.0001
+      scheduler: LinearWarmupCosineAnnealingLR
+      max_epochs: 50
+
+  - config: faenet-is2re_aux-all
+    note: "Cosine-low-lr"
+    optim:
+      lr_initial: 0.0005
+      scheduler: LinearWarmupCosineAnnealingLR
+      max_epochs: 50
+  
+  - config: faenet-is2re_aux-all
+    note: "Cosine-med-lr"
+    optim:
+      lr_initial: 0.001
+      scheduler: LinearWarmupCosineAnnealingLR
+      max_epochs: 50
+
+  - config: faenet-is2re_aux-all
+    note: "Cosine-usual-lr"
+    optim:
+      lr_initial: 0.002
+      scheduler: LinearWarmupCosineAnnealingLR
+      max_epochs: 50
+
+# default LR scheduler
+  - config: faenet-is2re_aux-all
+    note: "cst-very-low-lr"
+    optim:
+      lr_initial: 0.0001
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-all
+    note: "cst-low-lr"
+    optim:
+      lr_initial: 0.0005
+      max_epochs: 50
+      energy_coefficient: 0.0
+  
+  - config: faenet-is2re_aux-all
+    note: "cst-med-lr"
+    optim:
+      lr_initial: 0.001
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-all
+    note: "cst-usual-lr"
+    optim:
+      lr_initial: 0.002
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+# Null LR
+  - config: faenet-is2re_aux-all
+    note: "cst-very-low-lr"
+    optim:
+      lr_initial: 0.0001
+      scheduler: Null
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-all
+    note: "cst-low-lr"
+    optim:
+      lr_initial: 0.0005
+      scheduler: Null
+      max_epochs: 50
+      energy_coefficient: 0.0
+  
+  - config: faenet-is2re_aux-all
+    note: "cst-med-lr"
+    optim:
+      lr_initial: 0.001
+      scheduler: Null
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-all
+    note: "cst-usual-lr"
+    optim:
+      lr_initial: 0.002
+      scheduler: Null
+      max_epochs: 50
+      energy_coefficient: 0.0
+# ----------------------------10k-----------------------
+  - config: faenet-is2re_aux-10k
+    note: "Cosine-very-low-lr"
+    optim:
+      lr_initial: 0.0001
+      scheduler: LinearWarmupCosineAnnealingLR
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-10k
+    note: "Cosine-low-lr"
+    optim:
+      lr_initial: 0.0005
+      scheduler: LinearWarmupCosineAnnealingLR
+      max_epochs: 50
+      energy_coefficient: 0.0
+  
+  - config: faenet-is2re_aux-10k
+    note: "Cosine-med-lr"
+    optim:
+      lr_initial: 0.001
+      scheduler: LinearWarmupCosineAnnealingLR
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-10k
+    note: "Cosine-usual-lr"
+    optim:
+      lr_initial: 0.002
+      scheduler: LinearWarmupCosineAnnealingLR
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+# default LR scheduler
+  - config: faenet-is2re_aux-10k
+    note: "cst-very-low-lr"
+    optim:
+      lr_initial: 0.0001
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-10k
+    note: "cst-low-lr"
+    optim:
+      lr_initial: 0.0005
+      max_epochs: 50
+      energy_coefficient: 0.0
+  
+  - config: faenet-is2re_aux-10k
+    note: "cst-med-lr"
+    optim:
+      lr_initial: 0.001
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-10k
+    note: "cst-usual-lr"
+    optim:
+      lr_initial: 0.002
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+# Null LR
+  - config: faenet-is2re_aux-10k
+    note: "cst-very-low-lr"
+    optim:
+      lr_initial: 0.0001
+      scheduler: Null
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-10k
+    note: "cst-low-lr"
+    optim:
+      lr_initial: 0.0005
+      scheduler: Null
+      max_epochs: 50
+      energy_coefficient: 0.0
+  
+  - config: faenet-is2re_aux-10k
+    note: "cst-med-lr"
+    optim:
+      lr_initial: 0.001
+      scheduler: Null
+      max_epochs: 50
+      energy_coefficient: 0.0
+
+  - config: faenet-is2re_aux-10k
+    note: "cst-usual-lr"
+    optim:
+      lr_initial: 0.002
+      scheduler: Null
+      max_epochs: 50
+      energy_coefficient: 0.0
