+++ configs/models/faenet.yaml
@@ -108,10 +108,6 @@ is2re_aux:
     graph_rewiring: remove-tag-0
     frame_averaging: "2D" # {"2D", "3D", "DA", ""}
     fa_method: "se3-random" # {"", all, randon, det, se3-all, se3-randon, se3-det}
-    optim:
-      loss_position: "mse"
-      auxiliary_task_weight: 15
-      use_interpolate_init_relaxed_pos: True
     model:
       num_interactions: 18
       noisy_nodes: True
